Crate bandersnatch_vrfs

Bandersnatch thin VRF and ring VRFs

Re-exports

pub use ark_ed_on_bls12_381_bandersnatch as bandersnatch;

pub use dleq_vrf::transcript;

Modules

ring

vrf

VRF input and output handling

Structs

Message

PublicKey

RingVrfSignature

SecretKey

ThinVrfSignature

Transcript

Shake128 transcript style hasher.

Enums

SerializationError

This is an error that could occur during serialization

SignatureError

Constants

PUBLIC_KEY_LENGTH

Traits

CanonicalDeserialize

Deserializer in little endian format. This trait can be derived if all fields of a struct implement CanonicalDeserialize and the derive feature is enabled.

CanonicalSerialize

Serializer in little endian format. This trait can be derived if all fields of a struct implement CanonicalSerialize and the derive feature is enabled.

IntoTranscript

All types interpretable as Transcripts, including primarily impl BorrowMut<Traanscript> types like Transcript and &mut Transcript.

IntoVrfInput

Create VRF input points

Functions

pedersen_vrf

Pedersen VRF configured by the G1 generator for public key certs.

thin_vrf

Then VRF configured by the G1 generator for signatures.

Type Definitions

EdwardsAffine

PedersenVrfSignature

PublicKeyBytes

SignatureResult

VrfInOut

VrfInput

VrfPreOut

Derive Macros

CanonicalDeserialize

CanonicalSerialize