Module bandersnatch_vrfs::vrf

VRF input and output handling

We caution that ring VRFs based upon DLEQ proofs like ours require malleable pre-outputs, which become insecure if used in conjunction with “soft key derivation” ala BIP32.

Structs

VrfInOut

VRF input and pre-output paired together, possibly unverified.

VrfInput

Actual VRF input, consisting of an elliptic curve point.

VrfPreOut

VRF pre-output, possibly unverified.

Traits

IntoVrfInput

Create VRF input points

Functions

ark_hash_to_curve

attach_inputs_array

attach_inputs_vec

collect_preoutputs_array

collect_preoutputs_vec

vrfs_merge

Merge VRF input and pre-output pairs from the same signer, probably using variable time arithmetic