Function bandersnatch_vrfs::vrf::vrfs_merge

pub fn vrfs_merge<C, B>(t: &mut Transcript, ps: &[B]) -> VrfInOut<C>where
    C: AffineRepr,
    B: Borrow<VrfInOut<C>>,

Merge VRF input and pre-output pairs from the same signer, probably using variable time arithmetic

We merge VRF input and pre-outputs pairs by a single signer using the same technique as the “DLEQ Proofs” and “Batching the Proofs” sections of “Privacy Pass - The Math” by Alex Davidson, https://new.blog.cloudflare.com/privacy-pass-the-math/#dleqproofs and “Privacy Pass: Bypassing Internet Challenges Anonymously” by Alex Davidson, Ian Goldberg, Nick Sullivan, George Tankersley, and Filippo Valsorda. https://www.petsymposium.org/2018/files/papers/issue3/popets-2018-0026.pdf

As noted there, our merging technique’s soundness appeals to Theorem 3.17 on page 74 of Ryan Henry’s PhD thesis “Efficient Zero-Knowledge Proofs and Applications” https://uwspace.uwaterloo.ca/bitstream/handle/10012/8621/Henry_Ryan.pdf See also the attack on Peng and Bao’s batch proof protocol in “Batch Proofs of Partial Knowledge” by Ryan Henry and Ian Goldberg https://www.cypherpunks.ca/~iang/pubs/batchzkp-acns.pdf

We multiply every VrfInOut tuple here, which enables using faster 128 bit scalars. Amusingly, it turns out faster to do n 128 bit scalar multiplicaitons here, rather than merge these delinearization factors with the challenges used in signing and verifying.

…

We could reasonably ask if the VRF signer’s public key or the ring’s merkle root should be hashed when creating the scalars in vrfs_merge*, as Privacy Pass does. In principle, one could dicover relationships among the delinearizing scalars using k-sum attacks, but not among distinct VRF inputs because they’re hashed to the curve. TODO: Cite Wagner. We also note no such requirement when the values being hashed are BLS public keys as in https://crypto.stanford.edu/~dabo/pubs/papers/BLSmultisig.html